goInt, goUint and goBool typedefs for Go 1.1.
Go 1.1 allows implementations to make sizeof(int) = sizeof(int64) on
64 bit platforms: http://tip.golang.org/doc/go1.1#int.

It's not safe to assume that if we are targeting a 64 bit platform,
then the go implementation will use 64 bit size for int. Go 1.0 didn't
and while all the known implementations supporting Go 1.1 make use of
the full 64 bits, the size isn't forced by the spec, just allowed.

I'm not very familiar with qmake, but if the actual go implementation
could be probed for the int size, that would be ideal.